+++ src/gui.rs
@@ -257,7 +257,7 @@ impl eframe::App for GUI {
             let down_pressed = ui.ctx().input(|i| i.key_pressed(egui::Key::ArrowDown));
             if tab_pressed {
               ui.ctx().memory_mut(|mem| mem.request_focus(response.id));
-              let commands = vec![ "/clear","/reset","/persona ","/save ","/load ","/exit","/quit","/help"];
+              let commands = vec![ "/clear","/reset","/persona ","/scene ","/save ","/load ","/exit","/quit","/help"];
               if self.input.starts_with('/') {
                 let current_input = self.input.to_lowercase();
                 if current_input.starts_with("/persona ") {
@@ -373,6 +373,16 @@ impl eframe::App for GUI {
                       self.history_lines.push(("System".to_string(),format!("Please provide a valid persona: {}",self.config.personas.keys().cloned().collect::<Vec<String>>().join(","))));
                     }
                   },
+                  "/scene" => {
+                    if parts.len() > 1 {
+                      let scene: String = parts[1..].join(" ");
+                      self.chat.set_scene(&scene);
+                      self.history_lines.push(("System".to_string(),format!("Scene has been set to: {}",scene)));
+                      self.input.clear();
+                    } else {
+                      self.history_lines.push(("System".to_string(),format!("Please provide a scene.")));
+                    }
+                  },
                   "/save" => {
                     let filename: String = if parts.len() > 1 { parts[1..].join(" ") } else { self.config.history_file.clone() };
                     let msg = match self.chat.save_history(&filename.to_string(), &self.history_lines) {
@@ -405,6 +415,7 @@ impl eframe::App for GUI {
                        "/clear  - Clear the dispayed chat logs completely.",
                        "/reset  - Resets the current persona.",
                        format!("/persona [persona]  - Switches the active persona, 'persona' must be one of {}.",self.config.personas.keys().cloned().collect::<Vec<String>>().join(",")).as_str(),
+                       "/scene  - Sets the scene for the conversation, argument should be a single sentence.",
                        "/exit  - Safely close and exit the application.",
                        "/quit  - Also safely closes and exits the application.",
                        "/save [filename]  - Save session to a file (or configuration default).",